Refactor Stock entity and update database schema configuration
- Renamed table mapping for `Stock` entity from `stocks` to `vendor_medicine_price`.
- Updated `Stock` entity fields with specific column names: `quantity`, `expiry_date`, and changed `medicine_id` join column to `vendor_medicine_id`.
- Updated `MedicineRepository` native query to reflect the new table and column names (`vendor_medicine_price` and `vendor_medicine_id`).
- Cleaned up `application.properties` by removing redundant/commented-out database configurations and updating the active datasource URL.
- Changed `spring.jpa.hibernate.ddl-auto` to `none` and updated the Hibernate dialect to `MySQL8Dialect`.

+++ src/main/java/com/trego/dao/entity/OrderItem.java
@@ -21,10 +21,10 @@ public class OrderItem {
     private Order order;
 
     @Column(name = "medicine_id", nullable = false)
-    private Long medicineId;
+    private Integer medicineId;
 
     @Column(name = "vendor_medicine_id")
-    private Long vendorMedicineId;
+    private Integer vendorMedicineId;
 
     @Column(nullable = true)
     private Integer qty;
